components/Editor/Whisper/requirements.txt
torch
openai-whisper
moviepy==1.0.3
fastapi
uvicorn